Refactor by adding commands module
