+++ src/datapump/commands/set_latest.py
@@ -1,0 +1,10 @@
+from datapump.util.models import StrictBaseModel
+
+
+class SetLatestCommand(StrictBaseModel):
+    command: str
+
+    class SetLatestParameters(StrictBaseModel):
+        analysis_version: str
+
+    parameters: SetLatestParameters
