+++ src/lambdas/postprocessor/src/lambda_function.py
@@ -4,7 +4,7 @@ from typing import List, Union, cast
 from datapump.clients.aws import get_s3_client, get_s3_path_parts
 from datapump.clients.datapump_store import DatapumpConfig, DatapumpStore
 from datapump.clients.rw_api import update_area_statuses
-from datapump.commands import SyncType
+from datapump.commands.sync import SyncType
 from datapump.globals import GLOBALS, LOGGER
 from datapump.jobs.geotrellis import FireAlertsGeotrellisJob, GeotrellisJob
 from datapump.jobs.version_update import RasterVersionUpdateJob
