+++ tests/tests/test_unit.py
@@ -7,7 +7,7 @@ os.environ["S3_BUCKET_PIPELINE"] = "gfw-pipelines-test"
 os.environ["S3_BUCKET_DATA_LAKE"] = "gfw-data-lake-test"
 os.environ["GEOTRELLIS_JAR_PATH"] = "s3://gfw-pipelines-test/geotrellis/jars"
 
-from datapump.commands import Analysis, AnalysisInputTable
+from datapump.commands.analysis import Analysis, AnalysisInputTable
 from datapump.jobs.geotrellis import FireAlertsGeotrellisJob, JobStatus
